Update Fun_Create a dictionary with the name of students approved or reproved using their scores from a text file..py
Function that takes a text file with the students' scores of a complete subject, quizzes, assignments, midterm and final exam and create a dictionary with the name of each student and whether or not they passed the subject. Taking certain criteria:- The quizzes are 6 quizzes, the two with the worst grades are removed and the average the rest are equivalent to 25% of the grade.- There are 4 assignments, the one with the worst grade and the average are withdrawn of the remaining 3 are equal to 25%.- The midterm and the last exam are worth 25% each. - The student passes with >=60%